requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
langchain-chroma>=0.0.1
langchain-core>=0.1.0
pydantic>=2.5.2
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
chromadb>=0.4.18
peft>=0.5.0
trl>=0.7.1
huggingface-hub>=0.19.0
beautifulsoup4>=4.12.0
pdfminer.six>=20221105
schedule>=1.2.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
Pillow
datasets>=2.14.0
bitsandbytes>=0.41.0
einops>=0.6.0
protobuf>=3.20.0,<4.0.0
openai>=1.0.0
tiktoken>=0.4.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.0.1
